// smsf-masterdata-ui/smsf-xchange-ui/src/app/services/nc-rest/model/noncomplianceDTO.ts
/**
 * Automation API provider
 * Automation provides template of this description, and anyone may modify it in your own project.
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface NoncomplianceDTO { 
    action?: string;
    actionType?: number;
    answerType?: string;
    businessId?: string;
    closed?: number;
    comments?: string;
    controllerName?: string;
    createTime?: Date;
    description?: string;
    endTime?: Date;
    feedback?: string;
    feedbackTime?: Date;
    id?: string;
    interactionTime?: Date;
    ncTypeName?: string;
    nctypeId?: string;
    operatorId?: string;
    operatorName?: string;
    receiptId?: string;
    referenceNo?: string;
    rowNum?: number;
    startTime?: Date;
    statusId?: string;
}